optimized wheel picker with canvas hack
